requirements.txt
asgiref==3.6.0
Django==4.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
